Create README.md
Astra Coin is an experimental digital currency that enables instant, private payments to anyone, anywhere in the world. Astra Coin uses peer-to-peer technology to operate with no central authority: managing transactions and issuing money are carried out collectively by the network. Astra Coin Core is the name of the open source software which enables the use of this currency.